Move Info Endpoint
Move the information endpoint from the eval resource into its own
location. This endpoint currently only returns the python version but
can easily be expanded with more info later.
