+++ snekbox/api/resources/__init__.py
@@ -1,3 +1,7 @@
 from .eval import EvalResource
+from .info import InformationResource
 
-__all__ = ("EvalResource",)
+__all__ = (
+    "EvalResource",
+    "InformationResource",
+)
